Fix documentation for after and before
The results shown were incorrect and the example for `before` showed a
call to `after`. Also, the examples did not really make sense, since
(when corrected) they would insert a span inside an ul.

This replaces both examples with more real-world and correct examples
and also adds a short description in addition to the example.